refactor: Move commands to commands.py
Also allow `python -m goji` to use instead of binary
