+++ goji/__main__.py
@@ -1,0 +1,3 @@
+from goji.commands import cli
+
+cli()
